# colexec: add optimized implementation of LIKE '%foo%bar%' patterns