--- a/pkg/sql/colexec/colexecprojconst/proj_like_ops.eg.go
+++ b/pkg/sql/colexec/colexecprojconst/proj_like_ops.eg.go
@@ -251,6 +251,132 @@ func (p projContainsBytesBytesConstOp) Next() coldata.Batch {
 	return batch
 }
 
+type projSkeletonBytesBytesConstOp struct {
+	projConstOpBase
+	constArg [][]byte
+}
+
+func (p projSkeletonBytesBytesConstOp) Next() coldata.Batch {
+	batch := p.Input.Next()
+	n := batch.Length()
+	if n == 0 {
+		return coldata.ZeroBatch
+	}
+	vec := batch.ColVec(p.colIdx)
+	var col *coldata.Bytes
+	col = vec.Bytes()
+	projVec := batch.ColVec(p.outputIdx)
+	p.allocator.PerformOperation([]coldata.Vec{projVec}, func() {
+		// Capture col to force bounds check to work. See
+		// https://github.com/golang/go/issues/39756
+		col := col
+		projCol := projVec.Bool()
+		// Some operators can result in NULL with non-NULL inputs, like the JSON
+		// fetch value operator, ->. Therefore, _outNulls is defined to allow
+		// updating the output Nulls from within _ASSIGN functions when the result
+		// of a projection is Null.
+		_outNulls := projVec.Nulls()
+		if vec.Nulls().MaybeHasNulls() {
+			colNulls := vec.Nulls()
+			if sel := batch.Selection(); sel != nil {
+				sel = sel[:n]
+				for _, i := range sel {
+					if !colNulls.NullAt(i) {
+						// We only want to perform the projection operation if the value is not null.
+						arg := col.Get(i)
+
+						{
+							var idx, skeletonIdx int
+							for skeletonIdx < len(p.constArg) {
+								idx = bytes.Index(arg, p.constArg[skeletonIdx])
+								if idx < 0 {
+									break
+								}
+								arg = arg[idx+len(p.constArg[skeletonIdx]):]
+								skeletonIdx++
+							}
+							projCol[i] = skeletonIdx == len(p.constArg)
+						}
+					}
+				}
+			} else {
+				_ = projCol.Get(n - 1)
+				_ = col.Get(n - 1)
+				for i := 0; i < n; i++ {
+					if !colNulls.NullAt(i) {
+						// We only want to perform the projection operation if the value is not null.
+						arg := col.Get(i)
+
+						{
+							var idx, skeletonIdx int
+							for skeletonIdx < len(p.constArg) {
+								idx = bytes.Index(arg, p.constArg[skeletonIdx])
+								if idx < 0 {
+									break
+								}
+								arg = arg[idx+len(p.constArg[skeletonIdx]):]
+								skeletonIdx++
+							}
+							projCol[i] = skeletonIdx == len(p.constArg)
+						}
+					}
+				}
+			}
+			// _outNulls has been updated from within the _ASSIGN function to include
+			// any NULLs that resulted from the projection.
+			// If $hasNulls is true, union _outNulls with the set of input Nulls.
+			// If $hasNulls is false, then there are no input Nulls. _outNulls is
+			// projVec.Nulls() so there is no need to call projVec.SetNulls().
+			projVec.SetNulls(_outNulls.Or(*colNulls))
+		} else {
+			if sel := batch.Selection(); sel != nil {
+				sel = sel[:n]
+				for _, i := range sel {
+					arg := col.Get(i)
+
+					{
+						var idx, skeletonIdx int
+						for skeletonIdx < len(p.constArg) {
+							idx = bytes.Index(arg, p.constArg[skeletonIdx])
+							if idx < 0 {
+								break
+							}
+							arg = arg[idx+len(p.constArg[skeletonIdx]):]
+							skeletonIdx++
+						}
+						projCol[i] = skeletonIdx == len(p.constArg)
+					}
+				}
+			} else {
+				_ = projCol.Get(n - 1)
+				_ = col.Get(n - 1)
+				for i := 0; i < n; i++ {
+					arg := col.Get(i)
+
+					{
+						var idx, skeletonIdx int
+						for skeletonIdx < len(p.constArg) {
+							idx = bytes.Index(arg, p.constArg[skeletonIdx])
+							if idx < 0 {
+								break
+							}
+							arg = arg[idx+len(p.constArg[skeletonIdx]):]
+							skeletonIdx++
+						}
+						projCol[i] = skeletonIdx == len(p.constArg)
+					}
+				}
+			}
+			// _outNulls has been updated from within the _ASSIGN function to include
+			// any NULLs that resulted from the projection.
+			// If $hasNulls is true, union _outNulls with the set of input Nulls.
+			// If $hasNulls is false, then there are no input Nulls. _outNulls is
+			// projVec.Nulls() so there is no need to call projVec.SetNulls().
+		}
+	})
+	return batch
+}
+
 type projRegexpBytesBytesConstOp struct {
 	projConstOpBase
 	constArg *regexp.Regexp
@@ -551,6 +677,132 @@ func (p projNotContainsBytesBytesConstOp) Next() coldata.Batch {
 				for i := 0; i < n; i++ {
 					arg := col.Get(i)
 					projCol[i] = !bytes.Contains(arg, p.constArg)
+				}
+			}
+			// _outNulls has been updated from within the _ASSIGN function to include
+			// any NULLs that resulted from the projection.
+			// If $hasNulls is true, union _outNulls with the set of input Nulls.
+			// If $hasNulls is false, then there are no input Nulls. _outNulls is
+			// projVec.Nulls() so there is no need to call projVec.SetNulls().
+		}
+	})
+	return batch
+}
+
+type projNotSkeletonBytesBytesConstOp struct {
+	projConstOpBase
+	constArg [][]byte
+}
+
+func (p projNotSkeletonBytesBytesConstOp) Next() coldata.Batch {
+	batch := p.Input.Next()
+	n := batch.Length()
+	if n == 0 {
+		return coldata.ZeroBatch
+	}
+	vec := batch.ColVec(p.colIdx)
+	var col *coldata.Bytes
+	col = vec.Bytes()
+	projVec := batch.ColVec(p.outputIdx)
+	p.allocator.PerformOperation([]coldata.Vec{projVec}, func() {
+		// Capture col to force bounds check to work. See
+		// https://github.com/golang/go/issues/39756
+		col := col
+		projCol := projVec.Bool()
+		// Some operators can result in NULL with non-NULL inputs, like the JSON
+		// fetch value operator, ->. Therefore, _outNulls is defined to allow
+		// updating the output Nulls from within _ASSIGN functions when the result
+		// of a projection is Null.
+		_outNulls := projVec.Nulls()
+		if vec.Nulls().MaybeHasNulls() {
+			colNulls := vec.Nulls()
+			if sel := batch.Selection(); sel != nil {
+				sel = sel[:n]
+				for _, i := range sel {
+					if !colNulls.NullAt(i) {
+						// We only want to perform the projection operation if the value is not null.
+						arg := col.Get(i)
+
+						{
+							var idx, skeletonIdx int
+							for skeletonIdx < len(p.constArg) {
+								idx = bytes.Index(arg, p.constArg[skeletonIdx])
+								if idx < 0 {
+									break
+								}
+								arg = arg[idx+len(p.constArg[skeletonIdx]):]
+								skeletonIdx++
+							}
+							projCol[i] = skeletonIdx != len(p.constArg)
+						}
+					}
+				}
+			} else {
+				_ = projCol.Get(n - 1)
+				_ = col.Get(n - 1)
+				for i := 0; i < n; i++ {
+					if !colNulls.NullAt(i) {
+						// We only want to perform the projection operation if the value is not null.
+						arg := col.Get(i)
+
+						{
+							var idx, skeletonIdx int
+							for skeletonIdx < len(p.constArg) {
+								idx = bytes.Index(arg, p.constArg[skeletonIdx])
+								if idx < 0 {
+									break
+								}
+								arg = arg[idx+len(p.constArg[skeletonIdx]):]
+								skeletonIdx++
+							}
+							projCol[i] = skeletonIdx != len(p.constArg)
+						}
+					}
+				}
+			}
+			// _outNulls has been updated from within the _ASSIGN function to include
+			// any NULLs that resulted from the projection.
+			// If $hasNulls is true, union _outNulls with the set of input Nulls.
+			// If $hasNulls is false, then there are no input Nulls. _outNulls is
+			// projVec.Nulls() so there is no need to call projVec.SetNulls().
+			projVec.SetNulls(_outNulls.Or(*colNulls))
+		} else {
+			if sel := batch.Selection(); sel != nil {
+				sel = sel[:n]
+				for _, i := range sel {
+					arg := col.Get(i)
+
+					{
+						var idx, skeletonIdx int
+						for skeletonIdx < len(p.constArg) {
+							idx = bytes.Index(arg, p.constArg[skeletonIdx])
+							if idx < 0 {
+								break
+							}
+							arg = arg[idx+len(p.constArg[skeletonIdx]):]
+							skeletonIdx++
+						}
+						projCol[i] = skeletonIdx != len(p.constArg)
+					}
+				}
+			} else {
+				_ = projCol.Get(n - 1)
+				_ = col.Get(n - 1)
+				for i := 0; i < n; i++ {
+					arg := col.Get(i)
+
+					{
+						var idx, skeletonIdx int
+						for skeletonIdx < len(p.constArg) {
+							idx = bytes.Index(arg, p.constArg[skeletonIdx])
+							if idx < 0 {
+								break
+							}
+							arg = arg[idx+len(p.constArg[skeletonIdx]):]
+							skeletonIdx++
+						}
+						projCol[i] = skeletonIdx != len(p.constArg)
+					}
 				}
 			}
 			// _outNulls has been updated from within the _ASSIGN function to include