… pkg/sql/colexec/colexecsel/like_ops_test.go