This commit introduces optimized implementation of LIKE patterns of the
form `%foo%bar%` with any number of "skeleton" words in between. There
are two TPCH queries that use this pattern, and this commit speeds Q13
by about 2x on a single node on my laptop.

The logic for evaluating such pattern is that for each "skeleton" words
we find its first occurrence in the unprocessed part of the input
string. If it is not found, then the pattern doesn't match, if it is
found, then we advance the input string right past that first
occurrence.

Here is the comparison of the microbenchmarks of the optimized version
and the default unoptimized version on the same pattern:
```
BenchmarkLikeOps/selSkeletonBytesBytesConstOp-16       	   37070	     32054 ns/op	2044.58 MB/s
BenchmarkLikeOps/selRegexpSkeleton-16                  	    1797	    668154 ns/op	  98.09 MB/s
```

Release note: None