@mgartner 2022-04-19
*[pkg/sql/colexec/colexecsel/like_ops_test.go, line 95 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/80073#-N-xhcaw0wcP0JiaP4PY:-N02Cwgt7b9q_3IyHHRb:b-mraw0t) ([raw file](https://github.com/cockroachdb/cockroach/blob/c12d0443cc14d7f786d2c0632376ead6ac0bd076/pkg/sql/colexec/colexecsel/like_ops_test.go#L95)):*
Right, I wanted the test to cover the case in the conditional where the :skull: optimization would not apply.

> I started typing the code for the transformation here (in GetLikeOperatorType), but we probably should make the transformation in the optimizer. Thoughts?
 
Agreed that this would be a nice normalization rule. I created an issue: https://github.com/cockroachdb/cockroach/issues/80192


<!-- Sent from Reviewable.io -->